pyproject.toml
[project]
name = "image-uploader"
version = "0.1.0"
description = "Flask image uploader with compression and Cloudflare R2 storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "pillow>=10.0.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0"
]
